Remove duplicate logs when importing
Sometimes logs are uploaded more than once, are uploaded at halftime, or are
otherwise subsets of other logs. logs.tf is supposed to deal with this by
providing the updatelog parameter for uploaders, but sometimes it isn't used.
This is especially common when two different sources upload the same log.

We try to deduplicate our dataset by checking if different logs contain the
exact same round (except for the logid). If they do, then the log with the
greater logid (which was uploaded later) is chosen as the master copy, and any
earlier logs have all their data deleted, and have duplicate_of set to the
latest log's logid.

Signed-off-by: Sean Anderson <seanga2@gmail.com>