Cluster data by steamid64 first
This significantly improves performance. Most data is accessed by steamid, and
not logid. By clustering data by steamid64, we reduce the number of pages
queries for pages in /player/ need to read. This significantly reduces the
cold-cache load time.

player_name needs to be rewritten to be content-ful, since player_stats is now
WITHOUT ROWID.

Signed-off-by: Sean Anderson <seanga2@gmail.com>