Properly handle timezones in date filters
Previously, date filters were assumed to be in UTC. Because dates are displayed
in the user's timezone, this could result games played just before/after a
date filter being incorrectly displayed/filtered. This makes date filter
parsing timezone aware.

Unfortunately, there is no way to get the browser to create an ISO date string
including a timezene from the "date" form input, so we need to add the timezone
in javascript. This also adds a dependency on dateutils so we can parse IANA
timezones.

Ideally, we should really use a datetime-local input here, but it is unsupported
by firefox :l

Signed-off-by: Sean Anderson <seanga2@gmail.com>