Reorganize queries to avoid non-aggregate columns
Postgres is dumb and doesn't care if non-aggregate columns not in the group by
clause are from a table whose primary key *is* in the group-by clause.
Therefore, we have to restructure our queries to avoid those columns. The
primary way this is done is by moving group-by clauses into sub-selects (instead
of just joining the table and doing a group-by in the parent query). In one
instance, we just add another column to the group-by (though it is technically
unnecessary).

Signed-off-by: Sean Anderson <seanga2@gmail.com>