Use view instead of triggers for last_logid and friends
Postgres doesn't really like the sort of trigger we were using to keep
track of the last logid/nameid. After much trial and error, I have found a
suitable replacement in the form of a view+query. In addition to being
supported, it also doesn't require additional state. We still need the
vestigial player table so we can turn the lateral join into a view. In the
future, of course, this table might hold more information.

Signed-off-by: Sean Anderson <seanga2@gmail.com>