Handle NumericValueOutOfRange errors when parsing logs
Some logs have bogus data that we can't parse properly (such as huge
integers which won't fit in an INT). Instead of just dying, rollback and
continue with other logs. To support this, we now delete all logs in one
place, using a common table. We try to keep the log/log_json rows around if
possible for two reapons: first, we know not to try parsing this log again,
and second we can come back and try again later if we ever figure out a
good way to deal with errors.

Signed-off-by: Sean Anderson <seanga2@gmail.com>