Use temporary tables for inserting new logs
By keeping all our logs in temporary tables, we should significantly reduce
the performance penalty of deleting rows. In addition, we can now batch
inserts to do all at once. This might be important in the future if we
decide to use a staging database to do processing with before importing
sanitized data to Postgres.

Signed-off-by: Sean Anderson <seanga2@gmail.com>