Further optimizations to assigning formats
The original intent for this query was to do to something like

for logid in log:
    for steamid64 in player_stats:
       for classid in class_stats:
           sum(...)

but postgres doesn't figure out that it can use the steamid64 from player
stats to access class_stats via its primary key. So instead, just skip
player_stats and create an index so postgres does the right thing.

Signed-off-by: Sean Anderson <seanga2@gmail.com>