Split lesser-used columns of rounds into their own table
Most of the rounds columns are unused except for detecting duplicate
rounds. Move them to their own table.

We also had a problem where duplicates might not be detected properly if
the duplicate was in the same batch of logs. To fix this, UNION
log/round/round_extra with their public versions.

Signed-off-by: Sean Anderson <seanga2@gmail.com>