Add reference to logid from chat
We had a couple of console messages with NULL logids. Since the key was
NULL, then they don't have to reference player_stats. In turn, this meant
that they might reference a nonexistant log. Add an explicit constraint to
rectify this.

We can't just reference log and player separately, since then we have no
idea what name this message was created by.

Closes: #35
Signed-off-by: Sean Anderson <seanga2@gmail.com>