Add autocompletion for map names
One alternative considered was to query the server on every input event,
but as it turns out browsers can do the autocompletion as long as you give
them an appropriate list. Since I anticipate that most users won't ever
filter by map (and that the map list could double or triple some pages), we
lazily load it using javascript. Since this is an enhancement, I don't
think it matters too much.

Closes: #24
Signed-off-by: Sean Anderson <seanga2@gmail.com>