Fix navigation discarding filters
None of the solutions here is great, owing to a combination of factors.

* requests.args only includes GET parameters, and not route variables
* Just appending to url (or path, etc.) results in argument duplication
* Jinja doesn't have dictionary comprehensions
* Python doesn't support duplicate keyword arguments

So we need to re-add the route variables to request.args, since the
navigation macro wouldn't otherwise have the appropriate informatuion. We
do this by shoving both into kwargs because that's the easiest way to
create a new dict. Then we modify it with __setitem__ in a do block because
we can't use the normal `foo[bar] = baz` syntax. FWIW the navigation macro
did get simplified a little bit, but we introduced some additional
boilerplate to callers.

Closes: #43
Signed-off-by: Sean Anderson <seanga2@gmail.com>