Optimize leaderboard
Borrow from OLTP databases and create a "cube" view. Instead of
filtering/summing over round_wlt, just store those sums over everything we
want to filter. I don't know of a good way to filter by date for this, so
just leave that out.

We have to use sub-selects in our query to avoid postgres filtering after
the sequential scan over leaderboard_cube. However, we can't do this in the
query itself, since it causes indices to be skipped.

This also changes the display to remove the date filters (since they are
now useless). We also display the dating for better visibility into how we
sort.

Signed-off-by: Sean Anderson <seanga2@gmail.com>