Fix logs with offclasses being overcounted in leaderboard
In a return of a common bug, I forgot that something like

	JOIN class_stats (logid, steamid64)

will create a row for each class the player played that log. To avoid this,
only join on classes with at least 2/3 time.

Also don't create the view if it doesn't exist to (hopefully) avoid schema
loads taking far too long.

Fixes: 96f31e7 ("Optimize leaderboard")
Signed-off-by: Sean Anderson <seanga2@gmail.com>