Store rounds won/lost/tied in player_stats
The number one query on logs.tf at the moment (in terms of time) is the one
used in get_player to calculate w/l/t statistics. This query spends most of
its time going through round. This is because it must add up the result of
each round (in log_wlt). Unfortunately, round is ordered by logid, and
there is no easy way to order it by steamid64 (since all we have are
teamids).

To get around this, store each player's w/l/t stats in player_stats. This
around doubles the size of the table (due to all the duplication), but it
is the best way (IMO) to include this data. Every query which loads k/d/a
had also been using log_wlt, so this shouldn't actually result in too much
unnecessary data being loaded (since it would otherwise have to have been
calculated).

With this, we can finally dispense with the log_wlt view.

Signed-off-by: Sean Anderson <seanga2@gmail.com>