Switch to separate form for sorting
Instead of having individual links for each column to sort by, use a form
with a dropdown instead.  This is not terribly necessary for the
leaderboard, but it makes a lot more sense for pages with a lot of sorting
options. It also allows us to sort by things which aren't columns
themselves.

Signed-off-by: Sean Anderson <seanga2@gmail.com>