Remove GROUP BY clause from get_player query
Postgres incorrectly estimates the number of rows from a query like

	SELECT count(*)
	FROM foo
	WHERE bar = baz
	GROUP BY bar

as the number of rows selected by the where clause, instead of a single row
[1]. This can result in poor plans when this is a subquery, such as in
get_player. However, postgres correctly estimates the number of rows when
we remove the GROUP BY. Since this clause was redundant anyway, remove it.

[1] https://www.postgresql.org/message-id/16383-a560bff3cc8ed33c%40postgresql.org

Signed-off-by: Sean Anderson <seanga2@gmail.com>