Only filter out already-present logs if they are up-to-date
Logs may be updated. The log index includes the time the log was last
updated in its response. Using it, we can re-parse logs which have been
updated since we last imported them. To do this, we add an optional second
argument to the logid iterator. Then, in filter_logs, we also yield updated
logs.

Closes: #30
Signed-off-by: Sean Anderson <seanga2@gmail.com>