Support importing logs from clone_logs databases
clone_logs [1] is a project which has downloaded [2] most of the logs.tf
backlog. By importing from clone_logs databases, we can import logs much
more quickly than if we were doing so directly from logs.tf. Unfortunately,
the clone_logs schema flattens much of the structure present in logs.tf
jsons, which we retain in our database. So we need to do a lot of
reorganizing to keep things sane.  In addition, the clone_logs schema
renames a lot of columns to be different (more uniform). However, we need
to feed this data to our importer which expects logs.tf-style data, so we
need to undo all of that.

[1] https://github.com/ldesgoui/clone_logs
[2] https://mega.nz/#F!l9oGiKCb!lTWT2RSkTYv-TJZb92_ksA

Closes: #44
Signed-off-by: Sean Anderson <seanga2@gmail.com>