Skip logids more recent than our last request
When new logs are added, the first few logs in a new page may be the same
as the last few from the previous page. We try and adjust our offset to
account for this, but this only happens when requesting a new page. To
avoid downloading duplicate logs, keep track of the last logid we saw and
skip past it before parsing more logs.

Signed-off-by: Sean Anderson <seanga2@gmail.com>