Add name and avatar columns to player
The name column is so we don't have to look for a player's most recent log just to determine what
name to use for them. Until we import their current name, we start with the name used on their first
log. The avatar is of course so we can display an avatar. Not all players have avatars (notably
those who have deleted their accounts).

Signed-off-by: Sean Anderson <seanga2@gmail.com>