Convert K/D/A to /30 for trends
This converts the trends page to use K/D/A per 30 minutes instead of total.
This should help remove some of the noise with differing durations. In
addition, we also add handling for when we get no logs at all.  And now
that we have filters there is no need to do the CASE WHEN stuff for
duration, except for dt since not all logs have it.

Signed-off-by: Sean Anderson <seanga2@gmail.com>