Add logs page
This adds a page for logs. Multiple logs may be combined by specifying id
multiple times. The queries are pretty gnarly because we have to
reconstruct hierarchal data which got normalized upon insertion to the
database. However, we still get fairly good performance since we don't
actually have to read that much in the first place. This really is the area
where postgres shines.

Sorting is not supported. I think we will need to do it client side.

Closes: #7
Signed-off-by: Sean Anderson <seanga2@gmail.com>