Don't delete duplicate logs
This allows us greater flexibility when determining what to do with logs.
This may cause a bit of growth in the database (30% (!) of logs are dupes).
However, it's what we've been doing anyway since 5385109 ("Commit once per
log instead of batching"). If worst comes to worst, I can always go in and
delete the duplicates manually.

Signed-off-by: Sean Anderson <seanga2@gmail.com>