Reorganize python files
To facilitate deployment, I would like to have a python package for the
website. However, instead of packaging things as-is (with everything on the
top-level), I would like to take the opportunity to reorganize things a
bit. The new hierarchy is as follows:

	trends/ # Top-level modules common to both the site and importers
	trends/site/ # Modules related to the website itself
	trends/importer/ # Modules related to importing data

There are some minor edits to what is/isn't executable, since imports work differently now.
Unfortunately, it looks like if you want to run the importer without installing anything, you will
need to do something like

	python -m trends.importer ...

and similarly for the site itself

	FLASK_ENV=development python -m trends.site

Signed-off-by: Sean Anderson <seanga2@gmail.com>