importer: logs: Use scoreboard for score in A/D matches
In A/D matches, the round score is generally not related to who won the
match. Because the teams switch partially through the match, the winner of
the first round may lose the match. Further, the score of the last round is
generally 0-0, regardless of what the in-game scoreboard shows. For this
reason, use the log scoreboard, since that is the only consistent indicator
of winners.

This involves a rewrite of the query to push the round parts into a
subselect so we don't have to deal with shoving stuff into the group by (or
using max). We select the red/blue score at the last possible moment to
simplify the logic, and avoid duplicating expressions.

Examples: 3069780, 3070800, 3010081
Closes: #56
Signed-off-by: Sean Anderson <seanga2@gmail.com>