salt: Use a subdomain for netdata
This allows us to use client certificates to restrict access to netdata.
Now it can be accessed directly, so I don't have to mess around with socks
proxies (and proxying 127.0.0.1). We need to have 'unsafe-eval' in our CSP,
since jquery requires it for whatever reason (perhaps it is an old
version). To use this you will need to create a PKI as described in [1].
This also fixes the proxy_pass.

Fixes: cd2cf0a ("salt: Track metrics with netdata")
Signed-off-by: Sean Anderson <seanga2@gmail.com>
[1] https://fardog.io/blog/2017/12/30/client-side-certificate-authentication-with-nginx/