salt: Clone site into /srv/uwsgi/trends
This lets us remove a manual step from the server setup, since we can now
clone the source directly onto the server (instead of packaging it first).
We use an editable install so we can `git push` and still use a virtualenv.
When dependencies are updated, we will still need to do a pip install
manually (whereas before this would have been automated). There is some
funkiness going on with permissions, because salt assumes that an existing
directory is a git repo, and tries to run a command in it (which fails).

Signed-off-by: Sean Anderson <seanga2@gmail.com>