site: Add player filter
This adds the frontend to the player filter. We use tom-select since it has
no external dependencies, and seems fairly light. It was also (relatively)
easy to beat to my will.

For the moment, aliases and avatars are not displayed in search results.
However, I would like to add these. In addition, upon submission, the
players' names are lost.

Closes: #25
Signed-off-by: Sean Anderson <seanga2@gmail.com>