Add weapons importer
This adds an importer for weapon names. Weapons in TF2 logs correspond
primarily to kill icons.  TF2 has an "items_game.txt" file which has names
of items. We can use this to get most (80% or so) of the weapon names. The
rest are primarily kill methods with no item (taunts, spells, projectiles,
and sentries), default weapons (which don't have names defined in
items.txt), and miscellaneous weapons. I plan to also add a default naming
scheme which simply substitutes underscores with spaces and converts the
name to Title Case.

For now, names just live in a separate column in the weapons table. This is
not ideal in the long term, since I would like to combine separate
"weapons" with the same name.

Signed-off-by: Sean Anderson <seanga2@gmail.com>