salt: Add weapons import service
This imports new items when TF2 gets updated. The worst that happens here
is that we display an ugly name for a while, so we don't check very often.
We always import after a week so that if someone has the first log with
e.g.

	tf_weapon_foobar

then it gets a pretty name assigned.

Signed-off-by: Sean Anderson <seanga2@gmail.com>