Make teamid NOT NULL
This makes teamid NOT NULL, eliminating dummy rows for spectators. As a
consequence, chat now has a split foreign key, so that players do not have
to have played in the log to have chatted. This finally implements #35 for
real.

Temporary table foreign keys cannot reference real tables, so we can no
longer detect when a player is only present in chat by waiting for a
foreign key error. Work around this by always inserting into player. To
alleviate some of the impact of this, do nothing on conflict instead of
upserting.

I am experimenting with including the migrations for database changes.
Hopefully this will make it easier for those following along at home. In
the future these may be applied automatically, but for now they must be
applied manually.

Examples: 2844704, 2860321, 2788885, 3053999
Closes: #79
Signed-off-by: Sean Anderson <seanga2@gmail.com>