site: logs: Limit subqueries to avoid poor estimates
Postgresql lacks the ability to consider foreign keys when joining tables,
and always considers columsn in separate tables to be completely unrelated.
This can have disastrous results when those columns are *not*, in fact,
unrelated, but instead have a 1-to-1 or 1-to-many relationship. This can
cause postgres to underestimate the number of rows returned by a join by
over 1000x. Unfortunately, there is no way to correct this through the
creation of additional statistics. It is possible to use adaptive query
optimization [1] to address this, but it involves training on specific
queries and is not a general solution. Additionally, I don't think the
linked extension is ready for production use (judging by the quality of
their code, documentation, and comments in said documentation). An attempt
was made [2] to upstream some of these strategies, but it has stalled.

So for now, all we can do is limit the rows on one side of the join. This
ensures that a 1000x underestimate is only a LIMITx overestimate. Obviously
not ideal, but hopefully this will reduce query times.

[1] https://github.com/postgrespro/aqo
[2] https://commitfest.postgresql.org/35/2958/
Signed-off-by: Sean Anderson <seanga2@gmail.com>