site: log: Display players without any classes correctly
When a player has no classes, they will cause the log rendering to fail.
This is because tuples cannot be compared to None. Additionally, we cannot
iterate over None. Fix this by providing a suitable surrogate tuple, and by
not iterating when there are no classes. We also remove the "hider" class
in this instance to make it clear that that player cannot be exapanded.

Examples: 3027588
Signed-off-by: Sean Anderson <seanga2@gmail.com>