site: log: Fix subsequent pages not rendering
There is some kind of subtle limit/offset interaction going on with
subqueries here. I am not sure what causes it, but it is fixed by joining
to classes with LEFT JOIN. This is ok, especially in light of 1d18a7b
("site: log: Display players without any classes correctly"). I am just not
very sure why it doesn't work in the first place. With this change, we also
need to ensure that we don't render the class list if the player has no
classes.

Fixes: 35bf34f ("site: logs: Limit subqueries to avoid poor estimates")
Signed-off-by: Sean Anderson <seanga2@gmail.com>