salt: Switch to Munin for metrics
Netdata takes up a lot of memory/cpu all the time (around 100M or so). We
are already memory-constrained, so this is too much. This switches to
Munin. I'm not really happy with it, so I may switch to e.g.
prometheus+grafana instead. At the very least it is much lighter.

Signed-off-by: Sean Anderson <seanga2@gmail.com>