Add caching for player overview
Calculating the statistics for the player overview requires reading all the
rows in player_stats for that player. Try to mitigate this by caching the
results of the query for 5 minutes. We use memcached since it seems fairly
simple to deploy, and we don't need any more advanced reatures for the
moment.

Signed-off-by: Sean Anderson <seanga2@gmail.com>