schema: Remove primary_classid from player_stats
We can now calculate this information from classids and class_durations. Or
rather, we can calculate it relative to the other classes a player played.
I think this is likely good enough, outside of some strange logs with
negative duration.

To accomplish this, we create a view in place of the player stats table,
and rename the table to player_stats_backing. The view also contains some
commonly-used stats like class names and percentages.

Signed-off-by: Sean Anderson <seanga2@gmail.com>