Use materialized view for map popularity
Map suggestions are rarely requested, so they tend not to be in cache. They
also take a while to calculate, since we need to do a full scan of log.
However, they are not very time-sensitive (popularity does not change much
over the course of a day), so we can use a materialized view instead.

Signed-off-by: Sean Anderson <seanga2@gmail.com>