salt: db: Autovacuum more often
Vacuuming refreshes the visibility map. This greatly improves the performance of queries with
index-only scans. As such, it is important to run it regularly.

Much of the dataset is effectively read-only, while most of the churn
happens in the most recent hour of rows. This causes the autovacuum
threshold to be set too high with the default scale factor of 0.2 (since we
will never get to 20% dead tuples). Reduce the scaling factor so autovacuum
gets run.

Signed-off-by: Sean Anderson <seanga2@gmail.com>