players: Use class durations for events/30
By re-using the duration from the previous query we can avoid actually
having to join log/player_stats unless we are filtering by them. I think
this is incorrect for duplicate logs, but the original query was incorrect
in the first place (by using LEFT JOIN log_nodups). This improves
performance for what is effectively the front page of the site.

Signed-off-by: Sean Anderson <seanga2@gmail.com>