importer: Refactor fetch_players_logids
This refactors fetch_players_logids to use fewer variables and fix a bug.
The bug was that when new logs were added, we could still lose track of
logs (despite all the housekeeping trying to prevent that). Adopt a simpler
approach, where we only keep track of the offset and the total amount of
logs available. Previously-seen logs (or newly-added logs) will be skipped
by keeping track of the lowest logid seen so far. We also keep track of
seen logs directly, instead of relying on the results field. Overall this
is simpler and should be less buggy.

Signed-off-by: Sean Anderson <seanga2@gmail.com>