site: player: Fix 404 page for players with no logs
A query like

	SELECT sum(1) WHERE FALSE;

will always return one row, even if it is all NULLs. When a player has said
something in a match but not played (e.g. they only spectated) they can
have an entry in player but not player_stats. Because we CROSS JOIN onto
player, we would not yield a 404 error here. This causes later parts of the
page (which assume (e.g.) that winrate is non-null) to fail. Fix this by
ensuring that we are joining to something in the join condition.

Fixes: 8e0b859 ("Make teamid NOT NULL")
Signed-off-by: Sean Anderson <seanga2@gmail.com>