site: util: Properly handle a missing memcached
When we can't connect to the memcached server (such as unit tests, or when
it's just not running), we crash upon calling get(). This is because
creating a connection doesn't actually cause any network requests, so we
will never get an exception.

First, if there is no memcached server configured at all, don't bother
connecting in the first place. Second, make an immediate request for a
nonexistant key. This will cause the request to fail immediately. An
alternative strategy could be to wait for the caller to call get() (or
set()) and swap things out then. However, sometimes pylibmc returns
SomeErrors, which makes it difficult to disambiguate between connection
errors and other problems (which shouldn't be called). An extra call isn't
too bad, and is much simpler.

Fixes: df581ad ("Add caching for player overview")
Signed-off-by: Sean Anderson <seanga2@gmail.com>