Add integration tests
This adds some basic integration tests. I'm not too happy with the way
pytest does fixtures (at the same time both magical and rigid), so this may
be reworked. I would like to be able to swap in a real database and test on
a much large sample size as well. For the moment, I just have some logs
which have been problematic in the past (old, broken/missing properties,
etc.).

There are two big tests added here, and one small one. The first test
crawls the site and follows all the links it can find. The primary purpose
of this test is to ensure that I haven't introduced any broken links or
gross errors. I use a random selector to limit the amount of pages
crawled. Unfortunately, hidden form inputs are not filled in properly, so
no forms are crawled for now.

The second test crawls the site as well, but it fills in form parameters.
We use knowledge of data to guide hypothesis. I am not really happy with
the way the fixtures are set up here. I think it would be nice to be able
to create module-level strategies which can query the database, but AFIAK
this is very unweildy with fixtures. I would like to extend this test to
also exercise pagination (should be simple, we just don't have enough logs
to have multiple pages without setting limit) and ordering (I have no idea
where we would get the acceptable keys from). I'd also like to be able to
exclude 404 errors here as well. At the moment we allow them because of
spectators.

Finally, we have a bit more traditional testing for the search
functionality. Unlike most other pages on the site, it has a sort of API
(including /api) with predictable responses. I would like to have more
tests like this, but I don't know if it's possible for most pages.

Closes: #74
Signed-off-by: Sean Anderson <seanga2@gmail.com>