site: Set Last-Modified header
This sets the last modified header where possible. In general, we use the
logid time for the relevant page in order to determine when it was
modified. I would have preferred to integrate logs_last_modified with
get_logs somehow, but I don't think it's possible (at first glance).

For some pages, it is not possible to set this header. This is most at
issue for materialized views, which despite being semi-static have no easy
way to query when they were last refreshed. I have generally not set this
header for permanent redirects, since they should be cached anyway. The
search page could technically be updated any time we import new steam
names. However, I suspect that we won't get too many repeated searches in
the first place, so I have left it with the default cache control.

Signed-off-by: Sean Anderson <seanga2@gmail.com>