importer: logs: Improve deduplication performance
Postgresql still does a sequntial scan over rounds even when we limit it to
+-1 day. I think this is because postgers doesn't use the information that
time and logid are correlated. Because of this, it may think that we could
be looking at a different 2-day range for each log. Avoid this by
calculating a range of logids for all logs we are inserting. We do this in
a separate query to help out the planner (assuring it that the boundaries
are constant). Because we no longer need to use the time of each log, we
can avoid joining the log table. This allows us to use USING instead of ON,
because there are no duplicate rows (such as e.g. red_score) in both round
and log.

Closes: #66
Signed-off-by: Sean Anderson <seanga2@gmail.com>