salt: Enable backups
This adds backups using restic. They are stored on backblaze, but that is
primarily configured just through environmental variables. Perhaps in the
future some of this could be done with systemd credentials?

The database is 24 GiB at the moment, with 800k logs imported (around 25%
of the 3M total). The backup is around 16 GiB raw, and 6.4 GiB compressed.
I plan on keeping a 7 days + 4 weeks + 12 months of backups, which means we
will need to store (assuming no deduplication at all) around 150 GiB of
data. With all logs imported, that'd be around 550 GiB. I think this is
pretty reasonable, especially since we may have some additional space
savings.

Closes: #64
Signed-off-by: Sean Anderson <seanga2@gmail.com>