Partition log_json
Citus's columnar storage allows json to be compressed to around 50% of what
zstd can achieve. It does this by compressing multiple logs at once.
Additionally, the logs can be stored as JSON and not BYTEA, allowing
postgres's regular tools to be used. This will alleviate the need to have a
pythong script for every micration pulling new data from json. There are
some downsides:

- Rows cannot be updated or deleted from columnar tables
- Access to rows in columnar tables effectively requires a sequential scan

The latter is not an issue, since we only use log_json when doing
migrations, which would necessarily involve a sequential scan anyway. The
former is a bit of a problem, since logs created within the past ~day may
be updated. To get around this, we partition log_json into tables of around
100,000 logs each. This adds some additional maintenance related to
creating partitions. I looked into pg_partman, but it seemed to be about as
much effort to set up as to roll our own. Hopefully this will not be a
burden going forward.

Signed-off-by: Sean Anderson <seanga2@gmail.com>