importer: logs: Fix logs with explicit quickfix/vaccinator ubers
This correctly handles which list vaccinator and quickfix ubers separately.
Once again, derived values in parsed logs are totally bogus. In some cases,
'unknown' ubers are what the total amount of ubers should be. Fix this by
determining if known ubers equal unknown, and adjusting the total ubers if
that is the case.

Closes: #82
Signed-off-by: Sean Anderson <seanga2@gmail.com>