importer: Update demo importer
This is a follow-up to commit 400ca99 ("Add demo importer"). After
importing around 300,000 demos, I have a better idea about what queries are
efficient. We don't need a separate table for the demos, and we store the
list of players as an array row. The K/D/A turned out to be basically
useless as AFAICT demos.tf counts kills made after the round ends. We
weren't paying attention to players' teams anyway, so there's nothing lost
here. I'm keeping the rest of the columns, but I can't see using any of
them in the future except for mapid (and perhaps the url).

The migration script (aka mass linking) is a bit unusual. We can get good
performance from demos.players, but *only* if we unnest it (before
aggregating it again...). Not sure why this works, but it's faster than
using it directly. Perhaps it entices the query planner to create a
temporary table.

Signed-off-by: Sean Anderson <seanga2@gmail.com>