site: Add maps page
This page displays stats broken down by map. This was already a somewhat
possible by using the totals page, but I think this is a much nicer
interface. We use the same hierarchal format as used on the log page.
Because each row has the same number of columns, I added some helper arrows
on the left side to distiguish branch rows from leaf nodes. This is
inspired by Thunar's tree view.

Maps names are divided up into parts, using non-alphanumerics as
separators. Generally, the first part is the map type, the second is the
map, and the rest is some kind of version. Heuristicly, this works quite
well, as long as things like workshop prefixes are excluded. There are
still a lot of nonconforming (manually-entered) map names, so we collapse
them by default.

Signed-off-by: Sean Anderson <seanga2@gmail.com>