site: overview: Use temporary table for class stats
At the moment, postgresql severely underestimates the number of rows in
classes. I tried to address this previously by making it a materialized
subquery. However, we still underestimate twice (once when creating
classes, and once when joining to it). We can reduce this to once by making
classes a temporary table and running analyze on it.

Signed-off-by: Sean Anderson <seanga2@gmail.com>