site: overview: Replace averages with formats
I don't think the per-class averages are particularly interesting. Replace
them with format statistics, which should be useful for e.g. identifying
what formats someone plays. The per-class averages live on in the totals
page, which is a bit more about the trivial.

Hopefully this will also improve performance a bit (since we don't have to
read event_stats).

Signed-off-by: Sean Anderson <seanga2@gmail.com>