site: logs: Limit/filter earlier
Postgresql isn't smart enough to push down the filter/limit clauses to just
the log and player_stats part of the query, so do it ourselves. Previously,
it would fetch all rows matching the player in all tables, which was not
very efficient, especially since we usually just want a page (or less) of
logs. All the filter columns are in log/player_stats, so we can move those
clauses there unconditionally. However, the order clauses may depend on
stuff from hsg/hsr if the user is sorting by these columns. In that
(uncommon) case, just let the outer order/limit clause handle it.

Signed-off-by: Sean Anderson <seanga2@gmail.com>