Rework duplicate log calculation
This begins the work to start fixing how duplicates are calculated.
Currently, the logs which combine into another log are marked as
duplicates, not the combined log. We get much better data from uncombined
logs, so mark the combined stuff as duplicate. To support this, we are
switching to having duplicate_of be an array. This begins the process.

Signed-off-by: Sean Anderson <seanga2@gmail.com>