schema: Simplify null checks
There are several places where we want to ensure that some set of columns
are either all null or all not null. I had generally accomplished this by
doing

	CHECK ((foo ISNULL AND ...) OR (foo NOTNULL AND...))

but this is super verbose. We can condense things by doing

	CHECK ((foo ISNULL) = (bar ISNULL))

which is way shorter in cases of two columns. For three or more columns,
add a function which determines if all its arguments are equal, so we can
do

	CHECK (equal(foo ISNULL, ...))

Signed-off-by: Sean Anderson <seanga2@gmail.com>