Convert team to and enum
team was created as a table mostly because SQLite doesn't have enums.
However, it's tedious to join against, especially when making lots of
comparisons. I ended up assuming the primary key values anway. So just
convert it to an enum. This makes the log page a bit trickier, since I used
the primary key value for sorting, but we can get around that using a map.

Signed-off-by: Sean Anderson <seanga2@gmail.com>