importer: logs: Check for dupes with differing times
Although just checking for time/duration is fairly effective (90% or so),
it doesn't handle cases where the timestamps have been modified by the
combiner. We can catch more duplicates (98% or so) by comparing stats (like
we used to do before commit 42f46a2 ("importer: logs: Fix demo
deduplication"). This is done as a separate query to make writing the
conditions easier. It could be combined if there is adverse performance
impact.

Signed-off-by: Sean Anderson <seanga2@gmail.com>