Support importing ETF2L matches
This adds support for importing ETF2L match data. In the next few commits,
this data will be matched with logs and exposed through the interface. The
schema and whole import process have been designed (with pain) to be ready
for RGL support as well. This has caused the schema to be particularly
contorted, since RGL models its data in a significantly different way from
ETF2L. In many places, fields may be NULL for one league and NOT NULL for
another. This is controlled through a set of functions on the league enum.
I chose this method, since it allows for the regular foreign key
infrastructure, and will hopefully reduce rework (when compared to
open-coding the conditions).

The importer itself is done fairly differently from the logs importer.
Instead of going directly from the API json to importing it, we create an
intermediate representation which is then fed to the importer. This should
hopefully make it easier to paper over differences in the APIs. As usual,
the data is terrible, and the code already has several workarounds.

Hopefully this will make adding RGL support less awful, although I suspect
there will be problems to solve. No schema survives contact with the data.

Signed-off-by: Sean Anderson <seanga2@gmail.com>