Migrate steamid64 to playerid
Currently, we have around 100,000 players (2^17 or so), but we use a 64-bit
steamid64 to store them. This makes queries have to read more data, and
prevents us from using extensions like intarray. Migrate to using a 32-bit
surrogate key. The default value for this key is the lower 32-bits of the
steamid64, as the upper 64-bits are the same for all players except three
bots in one log. Just delete those players and their logs for now; I will
re-import the log later.

Signed-off-by: Sean Anderson <seanga2@gmail.com>