site: Convert to playerid
This converts the front-end to use playerid instead of steamid64. We still
use steamid64 for our urls and templates, so generally this conversion only
affects queries. For players' pages, we look up the playerid up front in
the same query where we determine last_active. In general, not too much
modification was required, as in many places we already joined to player
for avatarhash.

Signed-off-by: Sean Anderson <seanga2@gmail.com>