util: Allow arbitrary mapping of filter columns
The system of using prefixes to map filter columns to their database name
is not very extensible.  Additionally, picking only one column name makes
it difficult to add new queries which might use a different name for a
column when it is beign filtered. Convert to a mapping-based system where
columns can have an arbitrary name. Columns which do not need to be renamed
can be passed as varargs. While we're here, convert to f-strings.

Signed-off-by: Sean Anderson <seanga2@gmail.com>