site: league: Add matches page
Add a matches page. It lives in common because it will be reused for
competitions as well. Of minor interest here is division filtering. This
isn't used for leagues (as divisions are heterogenous), but it will be used
for competitions.

The SQL here has some materialized CTEs and lateral joins because it's the
only way to force postgres to perform properly. The map filtering is also
done in a nonstandard way so it actually functions as a filter (as opposed
to introducing nested sequential scans). Once again, another curse on the
postgres query planner.

Each match has a drop-down showing the individual logs. This will allow
easy evaluation of the linking algorithm. Right now, it appears to be
around 80% effective, so it needs some tweaking. There's no dedicated match
page. Instead, I just link directly to the logs.

Signed-off-by: Sean Anderson <seanga2@gmail.com>